Add BaseFragment
